MERGE SORT FILE UPLOADED LANGUAGE-C
The file contains the program of merge sort. The program follows the logic divide and conquer. c The mergearray( ) function is used for merging two halves. A function mergeSort( ) is used to divide the array which recursively calls itself and divide the array into two halves and the process repeats till the array cannot be divided. In simple words it divides input array in two halves, calls itself for the two halves and then merges the two sorted halves. The merge( ) function is used for merging two halves. And then the sorted array is displayed.