CREATED QUICK_SORT LANGUAGE-C
Quick Sort program file is created but the program is not able to sort. At the time it does not have any function(except main). The program takes an array as input and consider the last position(n-1) as pivot and sets all elements smaller than value at position pivot to the left of it and larger values at the right of it.
For example if input array is 5 3 4 2 1
than output is:- 1 5 3 4 2