SWAP FILE UPLOADED LANGUAGE-C
As this is a swapping program so the program takes two values as input. The value of x an the value of y. Where x and y are two variables. And then it swaps the values of x and y and gives the result. This program uses three variables x, y and temp. 
Logic:-
temp=x; // assigns value of x in temp
x=y; // assigns value of y in x
y=temp; // assign value of temp(which is the value of x) in y
 
In the next version we will use only two variables x and y.