SWAP PROGRAM UPDATED FOR TWO VARIABLES
In the previous version we used three variables in swapping. But this version uses only two variables for swapping.
Logic:-
x=x+y; // first it assigns the sum of x and y to x
y=x-y; // now x is the sum so if we subtract y from it then x will be left but we have to assign the value of x to y hence y=x-y
x=x-y; // now y contains the initial value of x so if we subtract it from sum the we will get the initial value of y but to swap it should be assigned to x hence x=x-y

So this way the value of two variables can be swapped  without the use of third variable.
In the next version we will find another way to swap.