SWAPPING USING XOR
In the previous version we saw who two values can be swapped using two variables. In this we have done some little changes. Now the program uses bit-wise XOR (Exclusive OR) to swap two number without use of third variable.
Logic:-
x=x^y;
y=x^y;
x=x^y;
XOR is a great way to save information, because you can remove it by doing an XOR again.
To understand this let's take three variables x, y and z.
Let z=x^y
Now z^x will be y and z^y will be x as it always gives the exclusive result.
In the main context we will get the value of x(initial) if we do this x^y(the value of x here is not initial value of x).
Now the value of y has become x(initial value) so if we do x=x^y which is equivalent to x=z^y (here y is the initial value of x) hence we got the initial value of y which get stores in x and hence the value has been swapped.
This is a cool trick, but don’t write this as an actual swap function. It has some limitation when we use it with pointers or address of objects.