QUICK SORT PROGRAM UPDATED LANGUAGE-C
In the previous program we had only one function that is partition( ) which does the sorting. The previous program was actually made without learning the program. It was made only by watching the logic of quick sort hence it is not same and also it does not have a quick-sort method. Also the method of partitioning array was not much efficient. This program has much changes in comparison with previous program although the logic is same. This program has three functions:-
1. partition( ) :- This functions partition the array according to pivot's value i.e. values last than pivot at the left side of pivot and values greater than pivot at the right side of pivot. The function return an int value which is the position of pivot after partitioning.
2. quickSort ( ):- As we know quick-sort is a divide and conquer algorithm so this function divides the problem in two sub problems according to the return value of partition( ).
3. swap( ) :- Oh we know this function take two values as argument and swaps them. The swap( ) is used in partitioning algorithm.

About the pivot :- In this program the pivot is taken as the last element of each sub-problem. We can take this pivot as a random number from the array elements and it is more faster and it achieves more faster sorting we will see it in the next version of this program.