STRINGREVERSE FILE UPLOADED LANGUAGE-C
This is a program in c language to reverse the given string. As we know stack is a LIFO(Last In First Out) data structure hence we can reverse a string by pushing the string character by character into the stack and then printing by popping characters from the string. In this program we have two functions (other than main) push( ) and pop( ):-
push( ):- This function pushes characters of the given string into the stack.
pop( ):- This function popes characters of the string which are printed the the order they popped.
This program is implemented using array representation of stack. Next we will try to improve this program by using linked stack.