CIRCULAR QUEUE PROGRAM UPLOADED LANGUAGE-C
In the previous program Queue.c we was having a problem of wastage of so much of memory. When we deQueue then each time a memory part is wasted and that cannot be used again. So we have used the concept of circular arrays in this program so that their is a proper utilization of memory.